fix(olt): OltOnuHistoryRecord model f64→f32 to match DB real/float4
DB columns (rx_power, tx_power, distance, temperature) are real/float4
but the sqlx::FromRow struct used Option<f64> → type mismatch on read.
Changed to Option<f32> to match the real column type, fixing the
'float8' error in ONU history tab.

+++ src-tauri/src/models/olt.rs
@@ -199,10 +199,10 @@ pub struct OltOnuHistoryRecord {
     pub mac: Option<String>,
     pub name: Option<String>,
     pub status: String,
-    pub rx_power: Option<f64>,
-    pub tx_power: Option<f64>,
-    pub distance: Option<f64>,
-    pub temperature: Option<f64>,
+    pub rx_power: Option<f32>,
+    pub tx_power: Option<f32>,
+    pub distance: Option<f32>,
+    pub temperature: Option<f32>,
     pub recorded_at: DateTime<Utc>,
 }
